* friends design:   - friend requests   - my friends   - invite a friend

+++ app/src/main/java/com/mobile/praaktishockey/ui/friends/view/FriendsFragment.kt
@@ -3,24 +3,22 @@ package com.mobile.praaktishockey.ui.friends.view
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
-import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.DividerItemDecoration
 import com.mobile.praaktishockey.R
-import com.mobile.praaktishockey.base.BaseFragment
-import com.mobile.praaktishockey.domain.common.InfoDialog
+import com.mobile.praaktishockey.base.temp.BaseFragment
+import com.mobile.praaktishockey.databinding.FragmentFriendsBinding
 import com.mobile.praaktishockey.domain.entities.FriendDTO
-import com.mobile.praaktishockey.domain.extension.getViewModel
-import com.mobile.praaktishockey.domain.extension.hide
-import com.mobile.praaktishockey.domain.extension.onClick
-import com.mobile.praaktishockey.domain.extension.show
+import com.mobile.praaktishockey.domain.extension.*
 import com.mobile.praaktishockey.ui.friends.adapter.FriendsAdapter
 import com.mobile.praaktishockey.ui.friends.adapter.FriendsItemListner
 import com.mobile.praaktishockey.ui.friends.vm.FriendsFragmentViewModel
 import kotlinx.android.synthetic.main.fragment_friends.*
 
-class FriendsFragment constructor(override val layoutId: Int = R.layout.fragment_friends) : BaseFragment(),
-    FriendsItemListner {
+class FriendsFragment constructor(override val layoutId: Int = R.layout.fragment_friends) :
+    BaseFragment<FragmentFriendsBinding>(), FriendsItemListner {
 
     companion object {
         val TAG = FriendsFragment::class.java.simpleName
@@ -32,7 +30,16 @@ class FriendsFragment constructor(override val layoutId: Int = R.layout.fragment
 
     override fun initUI(savedInstanceState: Bundle?) {
         initViewModel()
-        rvFriends.layoutManager = LinearLayoutManager(context!!)
+
+        binding.rvFriends.addItemDecoration(
+            DividerItemDecoration(
+                requireContext(),
+                DividerItemDecoration.VERTICAL
+            ).apply {
+                ContextCompat.getDrawable(requireContext(), R.drawable.list_divider)
+                    ?.let { setDrawable(it) }
+            }
+        )
     }
 
     private fun initViewModel() {
@@ -53,16 +60,18 @@ class FriendsFragment constructor(override val layoutId: Int = R.layout.fragment
     }
 
     override fun onDeleteFriendItem(friend: FriendDTO) {
-        val dialog = InfoDialog(context!!, String.format(
-            getString(R.string.are_you_sure_want_delete),
-            "${friend.friendFirstName} ${friend.friendLastName}?"
-        ),
-            object : InfoDialog.InfoDialogListener {
-                override fun onOkClicked() {
-                    mViewModel.deleteFriend(friend.friendEmail)
-                }
-            })
-        dialog.show()
+        materialAlert {
+            setMessage(
+                String.format(
+                    getString(R.string.are_you_sure_want_delete),
+                    "${friend.friendFirstName} ${friend.friendLastName}?"
+                )
+            )
+            setPositiveButton(R.string.ok) { dialog, which ->
+                mViewModel.deleteFriend(friend.friendEmail)
+            }
+            setNegativeButton(R.string.cancel) { dialog, which -> }
+        }?.show()
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
